Clean up sensor reading and reporting
