+++ project-autumn-pi/src/utils/ds18b20.ts
@@ -1,5 +1,9 @@
 import { sensorIdNameMap } from "../consts/ds18b20";
 import Sensor, { DS18B20, SensorData, SensorType } from "../types/Sensor";
+import { sendData } from "./utils";
+// Doesnt have types or use ESM :(
+// eslint-disable-next-line
+const ds18b20Raspi = require("ds18b20-raspi");
 
 export const getNameFromSensorId = (id: string): string | false => {
   const value = sensorIdNameMap.get(id);
@@ -10,7 +14,7 @@ export const getNameFromSensorId = (id: string): string | false => {
 export const transformDS18B20ToSensor = (
   ds18b20: DS18B20,
 ): Sensor => {
-  const { id } = ds18b20;
+  const { id, t } = ds18b20;
   const name = getNameFromSensorId(id);
 
   if (!name) throw Error(`No corresponding name for DS18B20 id: ${id}`);
@@ -19,6 +23,7 @@ export const transformDS18B20ToSensor = (
     id,
     name,
     type: SensorType.Temperature,
+    reading: { timestamp: Date.now(), value: t },
   };
 };
 
@@ -41,4 +46,15 @@ export const getDataForSensor = (
     value: ds18b20.t,
     timestamp: Date.now(),
   };
+};
+
+export const readAndPersistTemperatures: () => void = async () => {
+  const allDS18B20Sensors: DS18B20[] = ds18b20Raspi.readAllC();
+  const ds18b20Sensors: Sensor[] = transformDS18B20ArrayToSensorArray(
+    allDS18B20Sensors,
+  );
+
+  for (const sensor of ds18b20Sensors) {
+    await sendData(sensor);
+  }
 };
