+++ project-autumn-pi/src/utils/utils.ts
@@ -12,7 +12,7 @@ const getEnvironmentVariable: (key: EnvironmentVariable) => string = (key) => {
   return value;
 };
 
-export const sendData: (sensor: Sensor) => Promise<Response> = (sensor) => {
+export const sendData: (sensor: Sensor) => Promise<Response> = async (sensor) => {
   const endpoint = getEnvironmentVariable(EnvironmentVariable.API_ENDPOINT);
   const apiKey = getEnvironmentVariable(EnvironmentVariable.API_KEY);
 
@@ -20,5 +20,13 @@ export const sendData: (sensor: Sensor) => Promise<Response> = (sensor) => {
     apiKey,
     sensor,
   };
-  return fetch(`${endpoint}/postSaveData`, { method: "post", body: JSON.stringify(payload) });
+
+  const response = await fetch(`${endpoint}/postSaveData`, { method: "post", body: JSON.stringify(payload) });
+
+  if (!response.ok) {
+    const data: Error = JSON.parse(await response.text());
+    throw Error(data.message);
+  }
+
+  return response;
 };
